+++ src/CMakeLists.txt
@@ -27,7 +27,6 @@ set(HEADER_FILES
     CodeGen_C.h
     CodeGen_D3D12Compute_Dev.h
     CodeGen_GPU_Dev.h
-    CodeGen_GPU_Host.h
     CodeGen_Internal.h
     CodeGen_LLVM.h
     CodeGen_Metal_Dev.h
@@ -110,6 +109,7 @@ set(HEADER_FILES
     ModulusRemainder.h
     Monotonic.h
     ObjectInstanceRegistry.h
+    OffloadGPULoops.h
     OutputImageParam.h
     ParallelRVar.h
     Param.h
@@ -190,7 +190,6 @@ set(SOURCE_FILES
     CodeGen_C.cpp
     CodeGen_D3D12Compute_Dev.cpp
     CodeGen_GPU_Dev.cpp
-    CodeGen_GPU_Host.cpp
     CodeGen_Hexagon.cpp
     CodeGen_Internal.cpp
     CodeGen_LLVM.cpp
@@ -266,6 +265,7 @@ set(SOURCE_FILES
     ModulusRemainder.cpp
     Monotonic.cpp
     ObjectInstanceRegistry.cpp
+    OffloadGPULoops.cpp
     OutputImageParam.cpp
     ParallelRVar.cpp
     Parameter.cpp
